chore: clear compiler and lint warnings
- Mark pure projection types, HelpTopic, CaskFlowReleases and
  AppManagementPermission nonisolated — fixes all main-actor isolation
  warnings that would become errors in the Swift 6 language mode
- Split catalog grid/list/error views into ContentView+Catalog.swift,
  bringing ContentView under the 400-line lint cap
- Raise test-target file_length to 600: suites grow linearly with coverage
- Inline a constant to bring one test back under function_body_length
